+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Argument parsing modifier to parse an array of a given type: `+`
-- Implemented `sll_arg_state_t` and `sll_free_args`
+- Implemented `sll_arg_state_t`, `sll_free_args` and `sll_parse_args_list`
 - Object creation modifier to dereference array elements: `*`
 
 ### Changed